+++ examples/react-query/src/gen/hooks/useFindPetsByTagsInfiniteHook.ts
@@ -1,7 +1,7 @@
 import client from '@kubb/plugin-client/client'
 import type { FindPetsByTagsQueryResponse, FindPetsByTagsQueryParams, FindPetsByTags400 } from '../models/FindPetsByTags.ts'
 import type { RequestConfig, ResponseConfig } from '@kubb/plugin-client/client'
-import type { QueryKey, InfiniteQueryObserverOptions, UseInfiniteQueryResult } from '@tanstack/react-query'
+import type { InfiniteData, QueryKey, InfiniteQueryObserverOptions, UseInfiniteQueryResult } from '@tanstack/react-query'
 import { infiniteQueryOptions, useInfiniteQuery } from '@tanstack/react-query'
 
 export const findPetsByTagsInfiniteQueryKey = (params?: FindPetsByTagsQueryParams) => ['v5', { url: '/pet/findByTags' }, ...(params ? [params] : [])] as const
@@ -41,7 +41,7 @@ export function findPetsByTagsInfiniteQueryOptionsHook(params?: FindPetsByTagsQu
  * @link /pet/findByTags
  */
 export function useFindPetsByTagsInfiniteHook<
-  TData = ResponseConfig<FindPetsByTagsQueryResponse>,
+  TData = InfiniteData<ResponseConfig<FindPetsByTagsQueryResponse>>,
   TQueryData = ResponseConfig<FindPetsByTagsQueryResponse>,
   TQueryKey extends QueryKey = FindPetsByTagsInfiniteQueryKey,
 >(