+++ examples/solid-query/src/gen/hooks/createLoginUser.ts
@@ -1,8 +1,7 @@
 import client from '@kubb/plugin-client/client'
 import type { LoginUserQueryResponse, LoginUserQueryParams, LoginUser400 } from '../models/LoginUser.ts'
 import type { RequestConfig } from '@kubb/plugin-client/client'
-import type { QueryKey, CreateBaseQueryOptions, CreateQueryResult } from '@tanstack/solid-query'
-import { queryOptions, createQuery } from '@tanstack/solid-query'
+import { queryOptions } from '@tanstack/solid-query'
 
 export const loginUserQueryKey = (params?: LoginUserQueryParams) => [{ url: '/user/login' }, ...(params ? [params] : [])] as const
 
@@ -26,29 +25,4 @@ export function loginUserQueryOptions(params?: LoginUserQueryParams, config: Par
       return loginUser(params, config)
     },
   })
-}
-
-/**
- * @summary Logs user into the system
- * @link /user/login
- */
-export function createLoginUser<TData = LoginUserQueryResponse, TQueryData = LoginUserQueryResponse, TQueryKey extends QueryKey = LoginUserQueryKey>(
-  params?: LoginUserQueryParams,
-  options: {
-    query?: Partial<CreateBaseQueryOptions<LoginUserQueryResponse, LoginUser400, TData, TQueryData, TQueryKey>>
-    client?: Partial<RequestConfig>
-  } = {},
-) {
-  const { query: queryOptions, client: config = {} } = options ?? {}
-  const queryKey = queryOptions?.queryKey ?? loginUserQueryKey(params)
-  const query = createQuery(() => ({
-    ...(loginUserQueryOptions(params, config) as unknown as CreateBaseQueryOptions),
-    queryKey,
-    initialData: null,
-    ...(queryOptions as unknown as Omit<CreateBaseQueryOptions, 'queryKey'>),
-  })) as CreateQueryResult<TData, LoginUser400> & {
-    queryKey: TQueryKey
-  }
-  query.queryKey = queryKey as TQueryKey
-  return query
 }